Move getLatestMetadataLocation to IcebergUtil
